pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["api", "vault", "rotation", "workflow", "audit", "cli"]

[project]
name = "pam-lab"
version = "1.0.0"
description = "A lightweight Privileged Access Management (PAM) automation platform."
authors = [
    {name = "Raouf", email = "raouf@example.com"},
]
dependencies = [
    "fastapi==0.109.0",
    "uvicorn==0.27.0",
    "cryptography==42.0.0",
    "typer==0.12.0",
    "pydantic==2.6.0",
    "pydantic-settings==2.1.0",
    "pyyaml==6.0.1",
    "requests==2.31.0",
    "python-jose==3.3.0",
    "passlib==1.7.4",
    "python-multipart==0.0.7",
    "rich==13.7.0"
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
    "pytest-cov==4.1.0",
    "httpx==0.26.0",
    "ruff==0.2.1",
    "mypy==1.8.0",
    "types-requests",
    "types-PyYAML"
]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "ARG", "SIM"]
ignore = []

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = "test_*.py"

[tool.mypy]
plugins = [
  "pydantic.mypy"
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
